add validation to table-test-ed test suite
